requirements.txt
flask>=3.0
ultralytics>=8.1
pillow>=10.0
numpy>=1.24
opencv-python-headless>=4.8
